+++ src/commons_codec/transform/aws_dms.py
@@ -9,7 +9,14 @@ import typing as t
 import simplejson as json
 
 from commons_codec.exception import MessageFormatError, UnknownOperationError
-from commons_codec.model import ColumnType, ColumnTypeMapStore, PrimaryKeyStore, TableAddress
+from commons_codec.model import (
+    ColumnType,
+    ColumnTypeMapStore,
+    PrimaryKeyStore,
+    SQLOperation,
+    SQLParameterizedClause,
+    TableAddress,
+)
 
 logger = logging.getLogger(__name__)
 
@@ -66,35 +73,39 @@ class DMSTranslatorCrateDBRecord:
         self.primary_keys: t.List[str] = self.container.primary_keys[self.address]
         self.column_types: t.Dict[str, ColumnType] = self.container.column_types[self.address]
 
-    def to_sql(self) -> str:
+    def to_sql(self) -> SQLOperation:
         if self.operation == "create-table":
             pks = self.control.get("table-def", {}).get("primary-key")
             if pks:
                 self.primary_keys += pks
             # TODO: What about dropping tables first?
-            return f"CREATE TABLE IF NOT EXISTS {self.address.fqn} ({self.DATA_COLUMN} OBJECT(DYNAMIC));"
+            return SQLOperation(f"CREATE TABLE IF NOT EXISTS {self.address.fqn} ({self.DATA_COLUMN} OBJECT(DYNAMIC));")
 
         elif self.operation in ["load", "insert"]:
-            values_clause = self.record_to_values()
-            sql = f"INSERT INTO {self.address.fqn} ({self.DATA_COLUMN}) VALUES ('{values_clause}');"
+            record = self.record_to_values()
+            sql = f"INSERT INTO {self.address.fqn} ({self.DATA_COLUMN}) VALUES (:record);"
+            parameters = {"record": record}
 
         elif self.operation == "update":
-            values_clause = self.record_to_update()
+            record = self.record_to_values()
+            clause = self.update_clause()
             where_clause = self.keys_to_where()
-            sql = f"UPDATE {self.address.fqn} SET {values_clause} WHERE {where_clause};"
+            sql = f"UPDATE {self.address.fqn} SET {clause.set_clause} WHERE {where_clause};"
+            parameters = {"record": record}
 
         elif self.operation == "delete":
             where_clause = self.keys_to_where()
             sql = f"DELETE FROM {self.address.fqn} WHERE {where_clause};"
+            parameters = None
 
         else:
             message = f"Unknown CDC event operation: {self.operation}"
             logger.warning(message)
             raise UnknownOperationError(message, operation=self.operation, record=self.record)
 
-        return sql
+        return SQLOperation(sql, parameters)
 
-    def record_to_update(self) -> str:
+    def update_clause(self) -> SQLParameterizedClause:
         """
         Serializes an image to a comma-separated list of column/values pairs
         that can be used in the `SET` clause of an `UPDATE` statement.
@@ -106,17 +117,21 @@ class DMSTranslatorCrateDBRecord:
         OUT
         data['age'] = '33', data['attributes'] = '{"foo": "bar"}', data['name'] = 'John'
         """
-        constraints: t.List[str] = []
-        for column_name, column_value in self.record["data"].items():
+        clause = SQLParameterizedClause()
+        for column, value in self.record["data"].items():
             # Skip primary key columns, they cannot be updated
-            if column_name in self.primary_keys:
+            if column in self.primary_keys:
                 continue
 
-            constraint = f"{self.DATA_COLUMN}['{column_name}'] = '{column_value}'"
-            constraints.append(constraint)
-        return ", ".join(constraints)
+            param_name = f":{column}"
 
-    def record_to_values(self) -> str:
+            clause.columns.append(f"{self.DATA_COLUMN}['{column}']")
+            clause.names.append(param_name)
+            clause.values.append(value)  # noqa: PD011
+
+        return clause
+
+    def record_to_values(self) -> t.Dict[str, t.Any]:
         """
         Apply type translations to record, and serialize to JSON.
 
@@ -133,7 +148,7 @@ class DMSTranslatorCrateDBRecord:
                 if column_type is ColumnType.MAP and isinstance(value, str):
                     value = json.loads(value)
                 self.data[column_name] = value
-        return json.dumps(self.data)
+        return self.data
 
     def keys_to_where(self) -> str:
         """
@@ -169,7 +184,7 @@ class DMSTranslatorCrateDB:
         self.primary_keys = primary_keys or PrimaryKeyStore()
         self.column_types = column_types or ColumnTypeMapStore()
 
-    def to_sql(self, record: t.Dict[str, t.Any]) -> str:
+    def to_sql(self, record: t.Dict[str, t.Any]) -> SQLOperation:
         """
         Produce INSERT|UPDATE|DELETE SQL statement from load|insert|update|delete CDC event record.
         """